refactor(gemini): split monolithic gemini.go into focused files
gemini.go (702→205 lines, -71%): decompose into single-responsibility
files within the same package gemini:

- backend.go (182 lines): SDK init & backend resolution —
  initSDK, determineBackend, parseVertexAI, prepareAuthHeader,
  RefreshAuth, ResetConnections
- response.go (174 lines): response processing pipeline —
  SendChat, processResponse, checkGeminiTruncation,
  checkResponse, handleEmptyContent, handleNoCandidates
- request.go (145 lines): request preparation pipeline —
  prepareRequest, configureTools, configureThinking,
  applyThinkingBudget
- tools.go (48 lines): SDK tool schema conversion —
  toSDKTool, toSDKSchema
- gemini.go (205 lines): struct, constants, options, NewClient,
  GenerateImages

All methods remain on *Client receiver. No behavioral changes.
go test ./... passes all 65 packages including integration/e2e.
