+++ internal/infrastructure/llm/gemini/tools.go
@@ -1,0 +1,48 @@
+// Copyright (c) 2026 gosharplite@gmail.com
+// SPDX-License-Identifier: MIT
+
+package gemini
+
+import (
+	"github.com/gosharplite/tell-me-go/internal/domain/tools"
+	"google.golang.org/genai"
+)
+
+func toSDKTool(declarations []*tools.ToolDeclaration) []*genai.Tool {
+	if len(declarations) == 0 {
+		return nil
+	}
+	sdkDecls := make([]*genai.FunctionDeclaration, len(declarations))
+	for i, d := range declarations {
+		sdkDecls[i] = &genai.FunctionDeclaration{
+			Name:        d.Name,
+			Description: d.Description,
+			Parameters:  toSDKSchema(d.Parameters),
+		}
+	}
+	return []*genai.Tool{
+		{
+			FunctionDeclarations: sdkDecls,
+		},
+	}
+}
+
+func toSDKSchema(s *tools.Schema) *genai.Schema {
+	if s == nil {
+		return nil
+	}
+	res := &genai.Schema{
+		Type:        genai.Type(s.Type),
+		Description: s.Description,
+		Required:    s.Required,
+		Enum:        s.Enum,
+		Items:       toSDKSchema(s.Items),
+	}
+	if s.Properties != nil {
+		res.Properties = make(map[string]*genai.Schema)
+		for k, v := range s.Properties {
+			res.Properties[k] = toSDKSchema(v)
+		}
+	}
+	return res
+}
